Add jest, tests and split the files for them
