+++ src/lib/cache.js
@@ -1,0 +1,7 @@
+
+// Caching
+
+export default {
+  CSS: {},
+  springs: {}
+}
